Add text and LIKELY attributes

+++ Voortman3D/main.hpp
@@ -20,7 +20,7 @@ namespace Voortman3D {
 
 		const uint32_t randomVariableKey = 10;
 
-		double sawHeight{};
+		float sawHeight{};
 
 		VkClearColorValue backgroundColor = { 1.f, 1.f, 1.f, 1.f };
